Added switch on enum to handle application loop in main.c and game.c file to handle game loop
