+++ game/src/include/game.h
@@ -1,0 +1,4 @@
+// Declarations.
+void Init();
+void Tick(const float DELTA_TIME);
+void Draw();
